profile issues done

+++ Controller/portfolioController.js
@@ -1,5 +1,5 @@
 const Portfolio = require('../Model/portfolioModel')
-
+const fs = require('fs')
 /* LOAD PORTFOLIO PAGE */
 
 const loadPortfolio = async (req, res) => {
@@ -15,32 +15,49 @@ const loadPortfolio = async (req, res) => {
 
 const editPortfolio = async (req, res) => {
     try {
-
+        // console.log(req.file.filename,req.body);
         const existingPortfolio = await Portfolio.findOne();
         const updateObj = {};
-        if (req.file && req.file.filename) {
-            updateObj.images = req.file.filename;
-            if (existingPortfolio && existingPortfolio.images) {
-                fs.unlink(`Public/profile/${existingPortfolio.images}`, (err) => {
-                    if (err) {
-                        console.error("Error deleting previous image:", err);
-                    } else {
-                        console.log("Previous image deleted successfully");
-                    }
-                });
+        
+        if(!existingPortfolio){
+            const data = new Portfolio({
+                images: req.file.filename,
+                name:req.body.name,
+                about:req.body.about
+              });
+              await data.save();
+              return res.json({ success: true });
+        }else{
+
+            if (req.file && req.file.filename) {
+                updateObj.images = req.file.filename;
+                console.log("looo");
+                if (existingPortfolio && existingPortfolio.images) {
+                    fs.unlink(`Public/profile/${existingPortfolio.images}`, (err) => {
+                        if (err) {
+                            console.error("Error deleting previous image:", err);
+                        } else {
+                            console.log("Previous image deleted successfully");
+                        }
+                    });
+                }
             }
-        }
-        if (req.body.name) {
-            updateObj.name = req.body.name;
-        }
-        if (req.body.about) {
-            updateObj.about = req.body.about;
-        }
-        const result = await Portfolio.updateOne({}, { $set: updateObj });
-        if (result.modifiedCount > 0) {
-            return res.json({ success: true });
-        } else {
-            return res.json({ success: false });
+            if (req.body.name) {
+                updateObj.name = req.body.name;
+                console.log("dddd");
+            }
+            if (req.body.about) {
+                console.log("llll");
+                updateObj.about = req.body.about;
+            }
+            const result = await Portfolio.updateOne({}, { $set: updateObj });
+            console.log(result);
+            if (result.modifiedCount > 0) {
+                console.log(result);
+                return res.json({ success: true });
+            } else {
+                return res.json({ success: false });
+            }
         }
     } catch (error) {
         console.error(error);
